\name{DEG_table}
\alias{DEG_table}
\title{DEG Table}
\description{Generates a table of the differentially expressed genes.}
\usage{
table <- DEG_table(top_table, my_DEGs, save = FALSE, filename = "degtable.csv")
}
\arguments{
\item{top_table}{A dataframe with a row for the number top genes, as generated by topTable from the "limma" package or by make_toptable from degfind.}
\item{my_DEGs}{Character vector with names of differentially expressed genes, such as that generated by degfind's DEGs function.}
\item{save}{TRUE to save file. Default is FALSE.}
\item{filename}{Alternative name to save file as. Default is "degtable.csv"}
}
\value{
A tibble.
}
\examples{
library(degfind)
lm_stats <- make_lm(expression, study_design, "disease", "control")[[2]]
top_table <- make_toptable(lm_stats)
my_DEGs <- DEGs(lm_stats)
table <- DEG_table(top_table, my_DEGs, save = TRUE)
}
