\name{DEGs}
\alias{DEGs}
\title{DEGs}
\description{Generates a character vector with the differentially expressed genes.}
\usage{
my_DEGs <- DEGs(lm_stats, venn = TRUE, p.value = 0.05, lfc = 1, print = FALSE)
}
\arguments{
\item{lm_stats}{bayesian stats of class "MArrayLM", as generated by eBayes from the "limma" package, or the second element of the list generated by "make_lm" from this degfind package.}
\item{venn}{The default value is TRUE, to show the venn diagram.}
\item{p.value}{The default value is 0.05.}
\item{lfc}{The default value is 1.}
\item{print}{TRUE to print the number of DEGs in console. Default is FALSE.}
}
\value{
A character vector with the names of the differentially expressed gene's names.
}
\examples{
library(degfind)
lm_stats <- make_lm(expression, study_design, "disease", "control")[[2]]
my_DEGs <- DEGs(lm_stats)
}
