\name{make_toptable}
\alias{make_toptable}
\title{Make Toptable}
\description{Process Bayesian statistics of a linear model to produce a dataframe as limma's topTables, plus a ranking to account for low and high expression genes.}
\usage{
top_table <- make_toptable(lm_stats)
}
\arguments{
\item{lm_stats}{bayesian stats of class "MArrayLM", as generated by eBayes from the "limma" package, or the second element of the list generated by "make_lm" from this degfind package.}
}
\value{
A dataframe with a row for the number top genes, as generated by topTable from the "limma" package + "rank" column.
}
\examples{
lm_stats <- make_lm(expression, study_design, "L_UC", "control")[[2]]
top_table <- make_toptable(lm_stats)
}
