\name{volcano}
\alias{volcano}
\title{Volcano}
\description{Save volcano plot with differential expression data, where the x axis is the logFC and the y axis is the adjusted p value.}
\usage{
volcano(top_table, my_DEGs, plot = TRUE, filename = "myvolcano.png")
}
\arguments{
\item{top_table}{A dataframe with a row for the number top genes, as generated by topTable from the "limma" package or by make_toptable from degfind.}
\item{my_DEGs}{Character vector with names of differentially expressed genes, such as that generated by degfind's DEGs function.}
\item{plot}{TRUE to show plot. Default value = FALSE.}
\item{filename}{Alternative file name for saving the plot. The default is "volcano.png".}
}
\value{
A .png file with a volcano plot.
}
\examples{
lm_stats <- make_lm(expression, study_design, "disease", "control")[[2]]
top_table <- make_toptable(lm_stats)
my_DEGs <- DEGs(lm_stats)
volcano(top_table, my_DEGs)
}
